Inline Case files for CyclomaticComplexMethodSpec
